requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Werkzeug==2.3.7
python-dateutil==2.8.2
matplotlib==3.7.2
reportlab==4.0.4
openpyxl==3.1.2
python-dotenv==1.0.0
